simple PID

+++ moote_v3.1/code/motee_v3/Inc/motor.h
@@ -20,7 +20,7 @@
 
 
 void MOTOR_Init(void);
-void MOTOR_SetSpeed(uint16_t speed);
+void MOTOR_SetSpeed(float speed);
 void MOTOR_Enable(void);
 void MOTOR_Disable(void);
 void MOTOR_SetDirForward(void);
